requirements.txt
certifi==2020.12.5
chardet==3.0.4
dnspython==1.16.0
idna==2.10
lxml==4.6.2
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.1
mysql-connector-python==8.0.19
Pillow==8.1.0
protobuf==3.6.1
python-dateutil==2.8.1
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.9
urllib3==1.25.8
webdavclient3==3.14.1
